Move everything to interpreters
